+++ soaint-sgd-enterprise-artifacts/canonical-api-contract/src/main/java/co/com/soaint/foundation/canonical/correspondencia/CorReferidoDTO.java
@@ -4,6 +4,7 @@ import lombok.Builder;
 import lombok.Data;
 
 import javax.xml.bind.annotation.XmlRootElement;
+import java.math.BigInteger;
 
 /**
  * ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -19,12 +20,12 @@ import javax.xml.bind.annotation.XmlRootElement;
 @Builder(builderMethodName = "newInstance")
 @XmlRootElement(namespace = "http://soaint.com/domain-artifacts/cor-referido/1.0.0")
 public class CorReferidoDTO {
-    private Long ideReferido;
+    private BigInteger ideReferido;
     private String nroRadRef;
 
     public CorReferidoDTO(){super();}
 
-    public CorReferidoDTO(Long ideReferido, String nroRadRef){
+    public CorReferidoDTO(BigInteger ideReferido, String nroRadRef){
         this.ideReferido = ideReferido;
         this.nroRadRef = nroRadRef;
     }
